+++ src/molecules/user/elements.js
@@ -1,34 +1,47 @@
 import styled from 'styled-components/macro';
 
 /**
- * The user container creates a vertical-directed flexbox.
- * It also centers the content to connect the otherwise unrelated content visually.
+ * The user container defines some basic styling for all elements inside.
  */
 export const UserContainer = styled.div`
-	display: flex;
-	flex-direction: column;
-	align-items: center;
 	color: #6a737d;
 `;
 
 /**
- * The user avatar aligns the user's avatar with the name, using flexbox.
- * This technique is also applied for the description to center the user's name correctly.
+ * The user container meta element splits the screen in half and put the avatar and user in this half.
+ * It creates some emphasis on the user's name and also allows extra information without affecting this positioning.
+ */
+export const UserContainerMeta = styled.div`
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: flex-end;
+	height: 50vh;
+`;
+
+/**
+ * The user container info element centers the user's description horizontally.
+ * The length of this content does not affect the avatar and name positioning.
+ */
+export const UserContainerInfo = styled.div`
+	display: flex;
+	justify-content: center;
+`;
+
+/**
+ * The user avatar creates a slightly bigger avatar for the user.
+ * It's because this is the primary visual symbol of the user itself.
  */
 export const UserAvatar = styled.div`
-	display: flex;
-	flex: 1;
-	align-items: flex-end;
 	padding: 0.5rem;
 	font-size: 1.5em;
 `;
 
 /**
- * The user's name is also the title of the app and is therefore vertically centered.
- * It's positioned with flexbox to achieve this without knowing the sizes of both the avatar and description.
+ * The user's name is increased in size and styled to style the title of the page.
+ * It's because this is the actual human-identifiable information about the user.
  */
 export const UserName = styled.h1`
-	flex: 0;
 	margin: 0;
 	padding: 0.5rem;
 	text-align: center;
@@ -38,25 +51,19 @@ export const UserName = styled.h1`
 `;
 
 /**
- * The user description aligns the user's description with the name, using flexbox.
- * Because of this flexbox positioning, the size of the text has no impact on the alignment.
- * This technique is also applied for the avatar to center the name correctly.
+ * The user's description is styled to make it a bit more readable and to have a max width.
+ * This styling will create more visual coherence between the name and description.
  */
 export const UserDescription = styled.div`
-	display: flex;
-	flex: 1;
-	align-items: flex-start;
-	justify-content: center;
 	margin: 0;
 	padding: 0.5rem;
-	width: 100%;
 	max-width: 24em;
 	text-align: center;
 	line-height: 1.5;
 `;
 
 /**
- * The user description highlight styles a single word within this description.
+ * The user's description highlights styles keywords within this description.
  */
 export const UserDescriptionHighlight = styled.a`
 	color: #24292e;