+++ src/pages/app/app.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import GithubUser from 'src/organisms/github-user';
-import { AppContainer } from './elements';
 
 export default function AppPage() {
 	return (
-		<AppContainer>
+		<main role="main">
 			<GithubUser />
-		</AppContainer>
+		</main>
 	);
 }